requirements.txt
click==8.0.3
Flask==2.0.2
itsdangerous==2.0.1
Jinja2==3.0.2
kafka-python==2.0.2
MarkupSafe==2.0.1
numpy==1.21.2
opencv-contrib-python==4.5.3.56
Werkzeug==2.0.2
